# go-server/api/swagger.yaml
openapi: 3.0.0
info:
  title: OpenTerps
  description: An Open API that contains information about terpenes, the effects,
    and the cannabis varieties that contain them.
  termsOfService: https://wayhigh.we.bs
  contact:
    email: benjamminredden@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 0.0.2
...
servers:
- url: https://virtserver.swaggerhub.com/CountryFriedCoders/OpenTerps/0.0.1
...
paths:
...
  /cannabis:
    put:
      tags:
      - cannabis
      summary: Update an existing cannabis strain
      description: Update cannabis strain in the api
      operationId: updateCannabisStrain
      requestBody:
        $ref: '#/components/requestBodies/Cannabis'
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Cannabis strain not found
        "405":
          description: Validation exception
    post:
      tags:
      - cannabis
      summary: Add a new cannabis strains to the api
      description: Add new cannabis strains to the api
      operationId: addCannabisStrain
      requestBody:
        $ref: '#/components/requestBodies/Cannabis'
      responses:
        "405":
          description: Invalid input
  /cannabis/findByTerpene:
    get:
      tags:
      - cannabis
      summary: Finds cannabis strains by terpene
      description: Multiple status values can be provided with comma separated strings
      operationId: findCannabisStrainsByTerpene
      parameters:
      - name: terpene
        in: query
        description: Terpene values that need to be considered for filter
        required: true
        style: form
        explode: true
        schema:
          type: array
          items:
            type: string
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cannabis'
                x-content-type: application/xml
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Cannabis'
        "400":
          description: Invalid terpene value
  /cannabis/{cannabisId}:
    get:
      tags:
      - cannabis
      summary: Find cannabis strain by ID
      description: Returns a single cannabis strain
      operationId: getCannabisStrainById
      parameters:
      - name: cannabisId
        in: path
        description: ID of cannabis strain to return
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "200":
          description: successful operation
          content:
            application/xml:
              schema:
                $ref: '#/components/schemas/Cannabis'
            application/json:
              schema:
                $ref: '#/components/schemas/Cannabis'
        "400":
          description: Invalid ID supplied
        "404":
          description: Cannabis strain not found
    post:
      tags:
      - cannabis
      summary: Updates a cannabis strain in the api
      description: Update a single cannabis strain in the API
      operationId: updateCannabisStrainByID
      parameters:
      - name: cannabisId
        in: path
        description: ID of cannabis strain that needs to be updated
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/body_1'
      responses:
        "405":
          description: Invalid input
    delete:
      tags:
      - cannabis
      summary: Deletes a cannabis strain
      description: Deletes a cannabis strain from the API
      operationId: deleteCannabisStrain
      parameters:
      - name: api_key
        in: header
        required: true
        style: simple
        explode: false
        schema:
          type: string
      - name: cannabisId
        in: path
        description: Cannabis strain id to delete
        required: true
        style: simple
        explode: false
        schema:
          type: integer
          format: int64
      responses:
        "400":
          description: Invalid ID supplied
        "404":
          description: Cannabis strain not found
...
components:
  schemas:
...
    Category:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      example:
        name: name
        id: 6
      xml:
        name: Category
    Cannabis:
      type: object
      properties:
        id:
          type: integer
          format: int64
        strain:
          type: string
        terpenes:
          type: array
          xml:
            name: terpene
            wrapped: true
          items:
            $ref: '#/components/schemas/Terpene'
      example:
        strain: strain
        terpenes:
        - effects:
          - name: name
            id: 7
          - name: name
            id: 7
          tastes:
          - name: name
            id: 5
          - name: name
            id: 5
          smells:
          - name: name
            id: 5
          - name: name
            id: 5
          name: myrcene
          id: 0
          category:
            name: name
            id: 6
          strains:
          - name: name
            id: 2
          - name: name
            id: 2
          tags:
          - name: name
            id: 1
          - name: name
            id: 1
        - effects:
          - name: name
            id: 7
          - name: name
            id: 7
          tastes:
          - name: name
            id: 5
          - name: name
            id: 5
          smells:
          - name: name
            id: 5
          - name: name
            id: 5
          name: myrcene
          id: 0
          category:
            name: name
            id: 6
          strains:
          - name: name
            id: 2
          - name: name
            id: 2
          tags:
          - name: name
            id: 1
          - name: name
            id: 1
        id: 0
      xml:
        name: Cannabis
    Tag:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      example:
        name: name
        id: 1
      xml:
        name: Tag
    Effect:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      example:
        name: name
        id: 7
      xml:
        name: Effect
    Strain:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      example:
        name: name
        id: 2
      xml:
        name: Strain
    Taste:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      example:
        name: name
        id: 5
      xml:
        name: Taste
    Smell:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
      example:
        name: name
        id: 5
      xml:
        name: Smell
    Terpene:
      required:
      - name
      type: object
      properties:
        id:
          type: integer
          format: int64
        category:
          $ref: '#/components/schemas/Category'
        name:
          type: string
          example: myrcene
        tags:
          type: array
          xml:
            name: tag
            wrapped: true
          items:
            $ref: '#/components/schemas/Tag'
        tastes:
          type: array
          xml:
            name: taste
            wrapped: true
          items:
            $ref: '#/components/schemas/Taste'
        smells:
          type: array
          xml:
            name: smell
            wrapped: true
          items:
            $ref: '#/components/schemas/Smell'
        strains:
          type: array
          xml:
            name: strain
            wrapped: true
          items:
            $ref: '#/components/schemas/Strain'
        effects:
          type: array
          xml:
            name: effect
            wrapped: true
          items:
            $ref: '#/components/schemas/Effect'
      example:
        effects:
        - name: name
          id: 7
        - name: name
          id: 7
        tastes:
        - name: name
          id: 5
        - name: name
          id: 5
        smells:
        - name: name
          id: 5
        - name: name
          id: 5
        name: myrcene
        id: 0
        category:
          name: name
          id: 6
        strains:
        - name: name
          id: 2
        - name: name
          id: 2
        tags:
        - name: name
          id: 1
        - name: name
          id: 1
      xml:
        name: Terpene
...
    body_1:
      type: object
      properties:
        terpenes:
          type: array
          description: Updated terpenes of the cannabis strain
          items:
            $ref: '#/components/schemas/Terpene'
...
  requestBodies:
...
    Cannabis:
      description: Cannabis object
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Cannabis'
        application/xml:
          schema:
            $ref: '#/components/schemas/Cannabis'
      required: true
...