# go-server/go/model_taste.go
/*
 * OpenTerps
 *
 * An Open API that contains information about terpenes, the effects, and the cannabis varieties that contain them.
 *
 * API version: 0.0.2
 * Contact: benjamminredden@gmail.com
 * Generated by: Swagger Codegen (https://github.com/swagger-api/swagger-codegen.git)
 */
package swagger

type Taste struct {

	Id int64 `json:"id,omitempty"`

	Name string `json:"name,omitempty"`
}